+++ lib/p2p/packets/types/DisconnectingPacket.ts
@@ -1,22 +1,11 @@
 import Packet, { PacketDirection } from '../Packet';
 import PacketType from '../PacketType';
+import { DisconnectionReason } from '../../../types/enums';
 
 export type DisconnectingPacketBody = {
   reason: DisconnectionReason;
   payload?: string;
 };
-
-export enum DisconnectionReason {
-  ResponseStalling,
-  IncompatibleProtocolVersion,
-  UnexpectedIdentity,
-  ForbiddenIdentityUpdate,
-  ConnectedToSelf,
-  NotReadyForConnections,
-  Banned,
-  AlreadyConnected,
-  Shutdown,
-}
 
 class DisconnectingPacket extends Packet<DisconnectingPacketBody> {
   public get type() {
